Implementation of Dijkstras algorithm
Assumptions of input:
the weights are non-negative integers,
the names of the vertices are in the form "gi.j" for the corner vertex
in the top left of the grid cell in row i, column j (0≤i≤h and 0≤j≤w)
and "v.i.j.k" for the kth vertex in cell (i, j), where 0≤k<n,
each corner vertex has an edge to a non-corner vertex in each cell it
is adjacent to,
edges from corners vertices to interior vertices are listed with the
corner vertex first,
no edge is listed in the input twice,
each cell is connected,
the graph is connected, and
the source and destination vertices for shortest path queries exist and
are non-corner vertices in different cells.

Given this input it correctly computes shortest path from give source
to destination.